:boom: Unify the behavior of the --exclude option in the delete subcommand with that of the extract subcommand
