+++ cli/src/command/delete.rs
@@ -79,7 +79,9 @@ fn delete_file_from_archive(args: DeleteCommand) -> io::Result<()> {
             |entry| {
                 let entry = entry?;
                 let entry_path = entry.header().path();
-                if globs.matches_any(entry_path) && !exclude_globs.matches_any(entry_path) {
+                if globs.matches_any(entry_path)
+                    && !exclude_globs.starts_with_matches_any(entry_path)
+                {
                     return Ok(None);
                 }
                 Ok(Some(entry))
@@ -94,7 +96,9 @@ fn delete_file_from_archive(args: DeleteCommand) -> io::Result<()> {
             |entry| {
                 let entry = entry?;
                 let entry_path = entry.header().path();
-                if globs.matches_any(entry_path) && !exclude_globs.matches_any(entry_path) {
+                if globs.matches_any(entry_path)
+                    && !exclude_globs.starts_with_matches_any(entry_path)
+                {
                     return Ok(None);
                 }
                 Ok(Some(entry))
