+++ cli/tests/cli/delete.rs
@@ -1,3 +1,5 @@
+mod exclude;
+
 use crate::utils::{components_count, diff::diff, setup, TestResources};
 use clap::Parser;
 use portable_network_archive::{cli, command};
@@ -122,74 +124,6 @@ fn delete_output() {
     diff(
         concat!(env!("CARGO_TARGET_TMPDIR"), "/delete_output/in/"),
         concat!(env!("CARGO_TARGET_TMPDIR"), "/delete_output/out/"),
-    )
-    .unwrap();
-}
-
-#[test]
-fn delete_output_exclude() {
-    setup();
-    TestResources::extract_in(
-        "raw/",
-        concat!(env!("CARGO_TARGET_TMPDIR"), "/delete_output_exclude/in/"),
-    )
-    .unwrap();
-    command::entry(cli::Cli::parse_from([
-        "pna",
-        "--quiet",
-        "c",
-        concat!(
-            env!("CARGO_TARGET_TMPDIR"),
-            "/delete_output_exclude/delete_output_exclude.pna"
-        ),
-        "--overwrite",
-        concat!(env!("CARGO_TARGET_TMPDIR"), "/delete_output_exclude/in/"),
-    ]))
-    .unwrap();
-    command::entry(cli::Cli::parse_from([
-        "pna",
-        "--quiet",
-        "experimental",
-        "delete",
-        concat!(
-            env!("CARGO_TARGET_TMPDIR"),
-            "/delete_output_exclude/delete_output_exclude.pna"
-        ),
-        "**/raw/text.txt",
-        "--exclude",
-        "**/raw/**",
-        "--unstable",
-        "--output",
-        concat!(
-            env!("CARGO_TARGET_TMPDIR"),
-            "/delete_output_exclude/delete_excluded.pna"
-        ),
-    ]))
-    .unwrap();
-
-    command::entry(cli::Cli::parse_from([
-        "pna",
-        "--quiet",
-        "x",
-        concat!(
-            env!("CARGO_TARGET_TMPDIR"),
-            "/delete_output_exclude/delete_excluded.pna"
-        ),
-        "--overwrite",
-        "--out-dir",
-        concat!(env!("CARGO_TARGET_TMPDIR"), "/delete_output_exclude/out/"),
-        "--strip-components",
-        &components_count(concat!(
-            env!("CARGO_TARGET_TMPDIR"),
-            "/delete_output_exclude/in/"
-        ))
-        .to_string(),
-    ]))
-    .unwrap();
-
-    diff(
-        concat!(env!("CARGO_TARGET_TMPDIR"), "/delete_output_exclude/in/"),
-        concat!(env!("CARGO_TARGET_TMPDIR"), "/delete_output_exclude/out/"),
     )
     .unwrap();
 }
